presentations/2013-06/Python_Environment_Pt_2/client_2.py
print()

import dateutil
import requests

sesh = requests.session()
sesh.auth = ('username', 'password')
response = sesh.get('http://httpbin.org/get')

print('I did something great for my second client with requests and dateutil!\n')
