presentations/2013-06/Python_Environment_Pt_2/requirements.txt
distribute==0.6.34
python-dateutil==2.1
requests==1.2.3
six==1.3.0
